move confidential helper functions into tooling
