+++ dan_layer/engine/tests/confidential.rs
@@ -7,12 +7,15 @@ use tari_template_lib::{
     models::{Amount, ComponentAddress},
     prelude::ConfidentialOutputProof,
 };
+use tari_template_test_tooling::confidential::{
+    generate_confidential_proof, generate_withdraw_proof, generate_withdraw_proof_with_inputs,
+};
 use tari_template_test_tooling::{SubstateType, TemplateTest};
 use tari_transaction_manifest::ManifestValue;
 
-use self::utilities::*;
-
-fn setup(initial_supply: ConfidentialOutputProof) -> (TemplateTest, ComponentAddress, SubstateAddress) {
+fn setup(
+    initial_supply: ConfidentialOutputProof,
+) -> (TemplateTest, ComponentAddress, SubstateAddress) {
     let mut template_test = TemplateTest::new(vec![
         "tests/templates/confidential/faucet",
         "tests/templates/confidential/utilities",
@@ -36,9 +39,11 @@ fn mint_initial_commitment() {
     assert_eq!(total_supply, Amount(0));
 }
 
+#[allow(clippy::too_many_lines)]
 #[test]
 fn transfer_confidential_amounts_between_accounts() {
-    let (confidential_proof, faucet_mask, _change) = generate_confidential_proof(Amount(100_000), None);
+    let (confidential_proof, faucet_mask, _change) =
+        generate_confidential_proof(Amount(100_000), None);
     let (mut template_test, faucet, faucet_resx) = setup(confidential_proof);
 
     // Create an account
@@ -46,7 +51,8 @@ fn transfer_confidential_amounts_between_accounts() {
     let (account2, _owner2, _k) = template_test.create_owned_account();
 
     // Create proof for transfer
-    let proof = generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
+    let proof =
+        generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
 
     // Transfer faucet funds into account 1
     let vars = [
@@ -69,15 +75,39 @@ fn transfer_confidential_amounts_between_accounts() {
         .unwrap();
 
     let diff = result.finalize.result.expect("Failed to execute manifest");
-    assert_eq!(diff.up_iter().filter(|(addr, _)| *addr == account1).count(), 1);
-    assert_eq!(diff.down_iter().filter(|(addr, _)| *addr == account1).count(), 1);
-    assert_eq!(diff.up_iter().filter(|(addr, _)| *addr == faucet).count(), 1);
-    assert_eq!(diff.down_iter().filter(|(addr, _)| *addr == faucet).count(), 1);
+    assert_eq!(
+        diff.up_iter().filter(|(addr, _)| *addr == account1).count(),
+        1
+    );
+    assert_eq!(
+        diff.down_iter()
+            .filter(|(addr, _)| *addr == account1)
+            .count(),
+        1
+    );
+    assert_eq!(
+        diff.up_iter().filter(|(addr, _)| *addr == faucet).count(),
+        1
+    );
+    assert_eq!(
+        diff.down_iter().filter(|(addr, _)| *addr == faucet).count(),
+        1
+    );
     assert_eq!(diff.up_iter().count(), 5);
     assert_eq!(diff.down_iter().count(), 3);
 
-    let withdraw_proof = generate_withdraw_proof(&proof.output_mask, Amount(100), Some(Amount(900)), Amount(0));
-    let split_proof = generate_withdraw_proof(&withdraw_proof.output_mask, Amount(20), Some(Amount(80)), Amount(0));
+    let withdraw_proof = generate_withdraw_proof(
+        &proof.output_mask,
+        Amount(100),
+        Some(Amount(900)),
+        Amount(0),
+    );
+    let split_proof = generate_withdraw_proof(
+        &withdraw_proof.output_mask,
+        Amount(20),
+        Some(Amount(80)),
+        Amount(0),
+    );
 
     let vars = [
         ("faucet_resx", faucet_resx.into()),
@@ -87,7 +117,10 @@ fn transfer_confidential_amounts_between_accounts() {
             "withdraw_proof",
             ManifestValue::new_value(&withdraw_proof.proof).unwrap(),
         ),
-        ("split_proof", ManifestValue::new_value(&split_proof.proof).unwrap()),
+        (
+            "split_proof",
+            ManifestValue::new_value(&split_proof.proof).unwrap(),
+        ),
     ];
     let result = template_test
         .execute_and_commit_manifest(
@@ -110,24 +143,42 @@ fn transfer_confidential_amounts_between_accounts() {
         )
         .unwrap();
     let diff = result.finalize.result.expect("Failed to execute manifest");
-    assert_eq!(diff.up_iter().filter(|(addr, _)| *addr == account1).count(), 1);
-    assert_eq!(diff.down_iter().filter(|(addr, _)| *addr == account1).count(), 1);
-    assert_eq!(diff.up_iter().filter(|(addr, _)| *addr == account2).count(), 1);
-    assert_eq!(diff.down_iter().filter(|(addr, _)| *addr == account2).count(), 1);
+    assert_eq!(
+        diff.up_iter().filter(|(addr, _)| *addr == account1).count(),
+        1
+    );
+    assert_eq!(
+        diff.down_iter()
+            .filter(|(addr, _)| *addr == account1)
+            .count(),
+        1
+    );
+    assert_eq!(
+        diff.up_iter().filter(|(addr, _)| *addr == account2).count(),
+        1
+    );
+    assert_eq!(
+        diff.down_iter()
+            .filter(|(addr, _)| *addr == account2)
+            .count(),
+        1
+    );
     assert_eq!(diff.up_iter().count(), 5);
     assert_eq!(diff.down_iter().count(), 3);
 }
 
 #[test]
 fn transfer_confidential_fails_with_invalid_balance() {
-    let (confidential_proof, faucet_mask, _change) = generate_confidential_proof(Amount(100_000), None);
+    let (confidential_proof, faucet_mask, _change) =
+        generate_confidential_proof(Amount(100_000), None);
     let (mut template_test, faucet, _faucet_resx) = setup(confidential_proof);
 
     // Create an account
     let (account1, _owner1, _k) = template_test.create_owned_account();
 
     // Create proof for transfer
-    let proof = generate_withdraw_proof(&faucet_mask, Amount(1001), Some(Amount(99_000)), Amount(0));
+    let proof =
+        generate_withdraw_proof(&faucet_mask, Amount(1001), Some(Amount(99_000)), Amount(0));
 
     // Transfer faucet funds into account 1
     let vars = [
@@ -152,7 +203,8 @@ fn transfer_confidential_fails_with_invalid_balance() {
 
 #[test]
 fn reveal_confidential_and_transfer() {
-    let (confidential_proof, faucet_mask, _change) = generate_confidential_proof(Amount(100_000), None);
+    let (confidential_proof, faucet_mask, _change) =
+        generate_confidential_proof(Amount(100_000), None);
     let (mut template_test, faucet, faucet_resx) = setup(confidential_proof);
 
     // Create an account
@@ -161,12 +213,18 @@ fn reveal_confidential_and_transfer() {
 
     // Create proof for transfer
 
-    let proof = generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
+    let proof =
+        generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
     // Reveal 90 tokens and 10 confidentially
-    let reveal_proof = generate_withdraw_proof(&proof.output_mask, Amount(10), Some(Amount(900)), Amount(90));
+    let reveal_proof = generate_withdraw_proof(
+        &proof.output_mask,
+        Amount(10),
+        Some(Amount(900)),
+        Amount(90),
+    );
     // Then reveal the rest
     let reveal_bucket_proof =
-        generate_withdraw_proof(&reveal_proof.output_mask, Amount(0), Some(Amount(0)), Amount(10));
+        generate_withdraw_proof(&reveal_proof.output_mask, Amount(0), None, Amount(10));
 
     // Transfer faucet funds into account 1
     let vars = [
@@ -175,7 +233,10 @@ fn reveal_confidential_and_transfer() {
         ("account1", account1.into()),
         ("account2", account2.into()),
         ("proof", ManifestValue::new_value(&proof.proof).unwrap()),
-        ("reveal_proof", ManifestValue::new_value(&reveal_proof.proof).unwrap()),
+        (
+            "reveal_proof",
+            ManifestValue::new_value(&reveal_proof.proof).unwrap(),
+        ),
         (
             "reveal_bucket_proof",
             ManifestValue::new_value(&reveal_bucket_proof.proof).unwrap(),
@@ -215,18 +276,23 @@ fn reveal_confidential_and_transfer() {
         .unwrap();
 
     assert_eq!(
-        result.finalize.execution_results[12].decode::<Amount>().unwrap(),
+        result.finalize.execution_results[12]
+            .decode::<Amount>()
+            .unwrap(),
         Amount(10)
     );
     assert_eq!(
-        result.finalize.execution_results[13].decode::<Amount>().unwrap(),
+        result.finalize.execution_results[13]
+            .decode::<Amount>()
+            .unwrap(),
         Amount(90)
     );
 }
 
 #[test]
 fn attempt_to_reveal_with_unbalanced_proof() {
-    let (confidential_proof, faucet_mask, _change) = generate_confidential_proof(Amount(100_000), None);
+    let (confidential_proof, faucet_mask, _change) =
+        generate_confidential_proof(Amount(100_000), None);
     let (mut template_test, faucet, faucet_resx) = setup(confidential_proof);
 
     // Create an account
@@ -235,9 +301,15 @@ fn attempt_to_reveal_with_unbalanced_proof() {
 
     // Create proof for transfer
 
-    let proof = generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
+    let proof =
+        generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
     // Attempt to reveal more than input - change
-    let reveal_proof = generate_withdraw_proof(&proof.output_mask, Amount(0), Some(Amount(900)), Amount(110));
+    let reveal_proof = generate_withdraw_proof(
+        &proof.output_mask,
+        Amount(0),
+        Some(Amount(900)),
+        Amount(110),
+    );
 
     // Transfer faucet funds into account 1
     let vars = [
@@ -246,7 +318,10 @@ fn attempt_to_reveal_with_unbalanced_proof() {
         ("account1", account1.into()),
         ("account2", account2.into()),
         ("proof", ManifestValue::new_value(&proof.proof).unwrap()),
-        ("reveal_proof", ManifestValue::new_value(&reveal_proof.proof).unwrap()),
+        (
+            "reveal_proof",
+            ManifestValue::new_value(&reveal_proof.proof).unwrap(),
+        ),
     ];
 
     // TODO: Propagate error messages from runtime
@@ -279,7 +354,8 @@ fn attempt_to_reveal_with_unbalanced_proof() {
 
 #[test]
 fn multi_commitment_join() {
-    let (confidential_proof, faucet_mask, _change) = generate_confidential_proof(Amount(100_000), None);
+    let (confidential_proof, faucet_mask, _change) =
+        generate_confidential_proof(Amount(100_000), None);
     let (mut template_test, faucet, faucet_resx) = setup(confidential_proof);
 
     // Create an account
@@ -287,7 +363,8 @@ fn multi_commitment_join() {
 
     // Create proof for transfer
 
-    let withdraw_proof1 = generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
+    let withdraw_proof1 =
+        generate_withdraw_proof(&faucet_mask, Amount(1000), Some(Amount(99_000)), Amount(0));
     let withdraw_proof2 = generate_withdraw_proof(
         withdraw_proof1.change_mask.as_ref().unwrap(),
         Amount(1000),
@@ -317,7 +394,10 @@ fn multi_commitment_join() {
             "withdraw_proof2",
             ManifestValue::new_value(&withdraw_proof2.proof).unwrap(),
         ),
-        ("join_proof", ManifestValue::new_value(&join_proof.proof).unwrap()),
+        (
+            "join_proof",
+            ManifestValue::new_value(&join_proof.proof).unwrap(),
+        ),
     ];
     let result = template_test
         .execute_and_commit_manifest(
@@ -351,252 +431,22 @@ fn multi_commitment_join() {
         )
         .unwrap();
 
-    assert_eq!(result.finalize.execution_results[3].decode::<u32>().unwrap(), 1);
-    assert_eq!(result.finalize.execution_results[7].decode::<u32>().unwrap(), 2);
-    assert_eq!(result.finalize.execution_results[9].decode::<u32>().unwrap(), 1);
-}
-
-/// These would live in the wallet
-pub mod utilities {
-    use rand::rngs::OsRng;
-    use tari_common_types::types::{BulletRangeProof, PrivateKey, PublicKey, Signature};
-    use tari_crypto::{
-        commitment::{ExtensionDegree, HomomorphicCommitmentFactory},
-        errors::RangeProofError,
-        extended_range_proof::ExtendedRangeProofService,
-        keys::{PublicKey as _, SecretKey},
-        ristretto::bulletproofs_plus::{RistrettoExtendedMask, RistrettoExtendedWitness},
-        tari_utilities::ByteArray,
-    };
-    use tari_engine_types::confidential::{challenges, get_commitment_factory, get_range_proof_service};
-    use tari_template_lib::{
-        crypto::{BalanceProofSignature, RistrettoPublicKeyBytes},
-        models::{Amount, ConfidentialOutputProof, ConfidentialStatement, ConfidentialWithdrawProof, EncryptedData},
-    };
-
-    pub struct ConfidentialProofStatement {
-        pub amount: Amount,
-        pub mask: PrivateKey,
-        pub sender_public_nonce: PublicKey,
-        pub minimum_value_promise: u64,
-    }
-
-    pub fn generate_confidential_proof(
-        output_amount: Amount,
-        change: Option<Amount>,
-    ) -> (ConfidentialOutputProof, PrivateKey, Option<PrivateKey>) {
-        let mask = PrivateKey::random(&mut OsRng);
-        let output_statement = ConfidentialProofStatement {
-            amount: output_amount,
-            mask: mask.clone(),
-            sender_public_nonce: Default::default(),
-            minimum_value_promise: 0,
-        };
-
-        let change_mask = PrivateKey::random(&mut OsRng);
-        let change_statement = change.map(|amount| ConfidentialProofStatement {
-            amount,
-            mask: change_mask.clone(),
-            sender_public_nonce: Default::default(),
-            minimum_value_promise: 0,
-        });
-
-        let proof = generate_confidential_proof_from_statements(output_statement, change_statement).unwrap();
-        (proof, mask, change.map(|_| change_mask))
-    }
-
-    pub fn generate_balance_proof(
-        input_mask: &PrivateKey,
-        output_mask: &PrivateKey,
-        change_mask: Option<&PrivateKey>,
-        revealed_amount: Amount,
-    ) -> BalanceProofSignature {
-        let secret_excess = input_mask - output_mask - change_mask.unwrap_or(&PrivateKey::default());
-        let excess = PublicKey::from_secret_key(&secret_excess);
-        let (nonce, public_nonce) = PublicKey::random_keypair(&mut OsRng);
-        let challenge = challenges::confidential_withdraw64(&excess, &public_nonce, revealed_amount);
-
-        let sig = Signature::sign_raw_uniform(&secret_excess, nonce, &challenge).unwrap();
-        BalanceProofSignature::try_from_parts(sig.get_public_nonce().as_bytes(), sig.get_signature().as_bytes())
-            .unwrap()
-    }
-
-    pub struct WithdrawProofOutput {
-        pub output_mask: PrivateKey,
-        pub change_mask: Option<PrivateKey>,
-        pub proof: ConfidentialWithdrawProof,
-    }
-
-    pub fn generate_withdraw_proof(
-        input_mask: &PrivateKey,
-        output_amount: Amount,
-        change_amount: Option<Amount>,
-        revealed_amount: Amount,
-    ) -> WithdrawProofOutput {
-        let (output_proof, output_mask, change_mask) = generate_confidential_proof(output_amount, change_amount);
-        let total_amount = output_amount + change_amount.unwrap_or_else(Amount::zero) + revealed_amount;
-        let input_commitment = get_commitment_factory().commit_value(input_mask, total_amount.value() as u64);
-        let input_commitment = copy_fixed(input_commitment.as_bytes());
-        let balance_proof = generate_balance_proof(input_mask, &output_mask, change_mask.as_ref(), revealed_amount);
-
-        let output_statement = output_proof.output_statement;
-        let change_statement = output_proof.change_statement.unwrap();
-
-        WithdrawProofOutput {
-            output_mask,
-            change_mask,
-            proof: ConfidentialWithdrawProof {
-                inputs: vec![input_commitment],
-                output_proof: ConfidentialOutputProof {
-                    output_statement: ConfidentialStatement {
-                        commitment: output_statement.commitment,
-                        sender_public_nonce: Default::default(),
-                        encrypted_data: EncryptedData::default(),
-                        minimum_value_promise: output_statement.minimum_value_promise,
-                        revealed_amount,
-                    },
-                    change_statement: Some(ConfidentialStatement {
-                        commitment: change_statement.commitment,
-                        sender_public_nonce: Default::default(),
-                        encrypted_data: EncryptedData::default(),
-                        minimum_value_promise: change_statement.minimum_value_promise,
-                        revealed_amount: Amount::zero(),
-                    }),
-                    range_proof: output_proof.range_proof,
-                },
-                balance_proof,
-            },
-        }
-    }
-
-    pub fn generate_withdraw_proof_with_inputs(
-        input: &[(PrivateKey, Amount)],
-        output_amount: Amount,
-        change_amount: Option<Amount>,
-        revealed_amount: Amount,
-    ) -> WithdrawProofOutput {
-        let (output_proof, output_mask, change_mask) = generate_confidential_proof(output_amount, change_amount);
-        let input_commitments = input
-            .iter()
-            .map(|(input_mask, amount)| {
-                let input_commitment = get_commitment_factory().commit_value(input_mask, amount.value() as u64);
-                copy_fixed(input_commitment.as_bytes())
-            })
-            .collect();
-        let input_private_excess = input
-            .iter()
-            .fold(PrivateKey::default(), |acc, (input_mask, _)| acc + input_mask);
-        let balance_proof = generate_balance_proof(
-            &input_private_excess,
-            &output_mask,
-            change_mask.as_ref(),
-            revealed_amount,
-        );
-
-        let output_statement = output_proof.output_statement;
-        let change_statement = output_proof.change_statement;
-
-        WithdrawProofOutput {
-            output_mask,
-            change_mask,
-            proof: ConfidentialWithdrawProof {
-                inputs: input_commitments,
-                output_proof: ConfidentialOutputProof {
-                    output_statement: ConfidentialStatement {
-                        commitment: output_statement.commitment,
-                        // R and encrypted value are informational and can be left out as far as the VN is concerned
-                        sender_public_nonce: Default::default(),
-                        encrypted_data: Default::default(),
-                        minimum_value_promise: output_statement.minimum_value_promise,
-                        revealed_amount,
-                    },
-                    change_statement: change_statement.map(|change| ConfidentialStatement {
-                        commitment: change.commitment,
-                        sender_public_nonce: Default::default(),
-                        encrypted_data: Default::default(),
-                        minimum_value_promise: change.minimum_value_promise,
-                        revealed_amount: Amount::zero(),
-                    }),
-                    range_proof: output_proof.range_proof,
-                },
-                balance_proof,
-            },
-        }
-    }
-
-    fn copy_fixed<const SZ: usize>(bytes: &[u8]) -> [u8; SZ] {
-        let mut array = [0u8; SZ];
-        array.copy_from_slice(&bytes[..SZ]);
-        array
-    }
-
-    fn generate_confidential_proof_from_statements(
-        output_statement: ConfidentialProofStatement,
-        change_statement: Option<ConfidentialProofStatement>,
-    ) -> Result<ConfidentialOutputProof, RangeProofError> {
-        let proof_change_statement = change_statement.as_ref().map(|statement| ConfidentialStatement {
-            commitment: commitment_to_bytes(&statement.mask, statement.amount),
-            sender_public_nonce: RistrettoPublicKeyBytes::from_bytes(statement.sender_public_nonce.as_bytes())
-                .expect("[generate_confidential_proof] change nonce"),
-            encrypted_data: Default::default(),
-            minimum_value_promise: statement.minimum_value_promise,
-            revealed_amount: Amount::zero(),
-        });
-
-        let output_range_proof = generate_extended_bullet_proof(&output_statement, change_statement.as_ref())?;
-
-        Ok(ConfidentialOutputProof {
-            output_statement: ConfidentialStatement {
-                commitment: commitment_to_bytes(&output_statement.mask, output_statement.amount),
-                sender_public_nonce: RistrettoPublicKeyBytes::from_bytes(
-                    output_statement.sender_public_nonce.as_bytes(),
-                )
-                .expect("[generate_confidential_proof] output nonce"),
-                encrypted_data: Default::default(),
-                minimum_value_promise: output_statement.minimum_value_promise,
-                revealed_amount: Amount::zero(),
-            },
-            change_statement: proof_change_statement,
-            range_proof: output_range_proof.0,
-        })
-    }
-
-    fn generate_extended_bullet_proof(
-        output_statement: &ConfidentialProofStatement,
-        change_statement: Option<&ConfidentialProofStatement>,
-    ) -> Result<BulletRangeProof, RangeProofError> {
-        let mut extended_witnesses = vec![];
-
-        let extended_mask =
-            RistrettoExtendedMask::assign(ExtensionDegree::DefaultPedersen, vec![output_statement.mask.clone()])
-                .unwrap();
-
-        let mut agg_factor = 1;
-        extended_witnesses.push(RistrettoExtendedWitness {
-            mask: extended_mask,
-            value: output_statement.amount.value() as u64,
-            minimum_value_promise: output_statement.minimum_value_promise,
-        });
-        if let Some(stmt) = change_statement {
-            let extended_mask =
-                RistrettoExtendedMask::assign(ExtensionDegree::DefaultPedersen, vec![stmt.mask.clone()]).unwrap();
-            extended_witnesses.push(RistrettoExtendedWitness {
-                mask: extended_mask,
-                value: stmt.amount.value() as u64,
-                minimum_value_promise: stmt.minimum_value_promise,
-            });
-            agg_factor = 2;
-        }
-
-        let output_range_proof =
-            get_range_proof_service(agg_factor).construct_extended_proof(extended_witnesses, None)?;
-        Ok(BulletRangeProof(output_range_proof))
-    }
-
-    fn commitment_to_bytes(mask: &PrivateKey, amount: Amount) -> [u8; 32] {
-        let commitment = get_commitment_factory().commit_value(mask, amount.value() as u64);
-        let mut bytes = [0u8; 32];
-        bytes.copy_from_slice(commitment.as_bytes());
-        bytes
-    }
+    assert_eq!(
+        result.finalize.execution_results[3]
+            .decode::<u32>()
+            .unwrap(),
+        1
+    );
+    assert_eq!(
+        result.finalize.execution_results[7]
+            .decode::<u32>()
+            .unwrap(),
+        2
+    );
+    assert_eq!(
+        result.finalize.execution_results[9]
+            .decode::<u32>()
+            .unwrap(),
+        1
+    );
 }
