+++ dan_layer/template_test_tooling/Cargo.toml
@@ -11,6 +11,7 @@ license = "BSD-3-Clause"
 [dependencies]
 tari_dan_engine = { path = "../engine" }
 tari_crypto = "0.19"
+tari_common_types = { git = "https://github.com/tari-project/tari.git", branch = "feature-dan2" }
 tari_template_lib = { path = "../template_lib" }
 tari_engine_types = { path = "../engine_types" }
 tari_transaction_manifest = { path = "../transaction_manifest" }
