+++ dan_layer/template_test_tooling/src/support/mod.rs
@@ -2,3 +2,4 @@
 //   SPDX-License-Identifier: BSD-3-Clause
 
 pub mod assert_error;
+pub mod confidential;
